backend/requirements.txt
fastapi==0.92.0
uvicorn==0.20.0
sqlalchemy==2.0.4
alembic==1.9.4
psycopg2-binary==2.9.5
boto3==1.26.91
python-multipart==0.0.6
python-dotenv==1.0.0